refactor: fixed issues in sponsor update
